refactor(db): extract checkpoint sink protocol
